changed int to float to avoid warnings fixed bug in level.cpp fixed bug with backspace in input name form fixed bug with incorrect players scores in records.cpp tested, works well next step - use shared_ptr in all cases it has possible

+++ SnakeGameSolution/SnakeGame/Block.cpp
@@ -4,7 +4,7 @@
 
 
 // constructor
-Block::Block(std::string BlockType, int x, int y, float XScale, float YScale)
+Block::Block(std::string BlockType, float x, float y, float XScale, float YScale)
 {
 	// set start block position on the game screen
 	MainSprite.setPosition(x, y);
